# openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "DevSecTools",
    "version": "1",
    "description": "DevSecTools scans websites for security-related information, and includes additional tools which streamline developer productivity.\n\nIt exposes a JSON-over-HTTP interface. Both `GET` and `POST` requests perform the same action.",
    "contact": {
      "name": "DevSecTools API",
      "url": "https://devsec.tools",
      "email": "api@devsec.tools"
    },
    "termsOfService": "https://devsec.tools/terms-of-service",
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/license/apache-2-0"
    }
  },
  "servers": [
    {
      "url": "https://api.devsec.tools",
      "description": "Production"
    },
    {
      "url": "http://api.devsec.local",
      "description": "Localdev"
    }
  ],
  "paths": {
    "/domain": {
      "get": {
        "summary": "GET domain",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/DomainResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "get-domain",
        "x-stoplight": {
          "id": "f16rgrfm8zgj5"
        },
        "security": [],
        "description": "Parses the hostname (domain) from the input.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StandardRequest"
          }
        ]
      },
      "post": {
        "summary": "POST domain",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/DomainResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "post-domain",
        "x-stoplight": {
          "id": "uyewgwqaxlypn"
        },
        "description": "Parses the hostname (domain) from the input.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/StandardRequest"
        }
      }
    },
    "/http": {
      "get": {
        "summary": "GET http",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HttpResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "get-http",
        "x-stoplight": {
          "id": "fphqqkvsw0uis"
        },
        "description": "Scans the hostname (domain) for HTTP/1.1, HTTP/2, and HTTP/3 support.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/StandardRequest"
          }
        ]
      },
      "post": {
        "summary": "POST http",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HttpResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "post-http",
        "x-stoplight": {
          "id": "gvqf6zpb9sn57"
        },
        "description": "Scans the hostname (domain) for HTTP/1.1, HTTP/2, and HTTP/3 support.",
        "requestBody": {
          "$ref": "#/components/requestBodies/StandardRequest"
        }
      }
    },
    "/tls": {
      "get": {
        "summary": "GET tls",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TlsResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "get-tls",
        "x-stoplight": {
          "id": "686gi3f7xgqk5"
        },
        "description": "Scans the hostname (domain) for supported TLS versions, as well as cipher suites per TLS version.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StandardRequest"
          }
        ]
      },
      "post": {
        "summary": "POST tls",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TlsResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse400"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse500"
          }
        },
        "operationId": "post-tls",
        "x-stoplight": {
          "id": "8c8uc6pvofyet"
        },
        "description": "Scans the hostname (domain) for supported TLS versions, as well as cipher suites per TLS version.   ",
        "requestBody": {
          "$ref": "#/components/requestBodies/StandardRequest"
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DomainResponse": {
        "type": "object",
        "x-stoplight": {
          "id": "e051c7e3c4e49"
        },
        "x-examples": {
          "Hostname": {
            "hostname": "www.apple.com"
          }
        },
        "description": "The response from a call to the `/domain` endpoint.",
        "title": "DomainResponse",
        "required": [
          "hostname"
        ],
        "properties": {
          "hostname": {
            "type": "string",
            "x-stoplight": {
              "id": "4eg7rp9ibo3rd"
            },
            "description": "The hostname (domain) that was parsed from the input."
          }
        },
        "examples": [
          {
            "hostname": "example.com"
          }
        ]
      },
      "HttpResponse": {
        "type": "object",
        "x-examples": {
          "HTTP Response": {
            "hostname": "https://stripe.com",
            "http11": true,
            "http2": true,
            "http3": false
          }
        },
        "description": "The response from a call to the `/http` endpoint.",
        "required": [
          "hostname",
          "http11",
          "http2",
          "http3"
        ],
        "properties": {
          "hostname": {
            "type": "string",
            "description": "The hostname (domain) that was parsed from the input."
          },
          "http11": {
            "type": "boolean",
            "description": "Whether or not the site supports HTTP/1.1 connections.\n\nA value of `true` means that the site supports HTTP/1.1 connections. A value of `false` means that it does not."
          },
          "http2": {
            "type": "boolean",
            "description": "Whether or not the site supports HTTP/2 connections.\n\nA value of `true` means that the site supports HTTP/2 connections. A value of `false` means that it does not."
          },
          "http3": {
            "type": "boolean",
            "description": "Whether or not the site supports HTTP/3 connections.\n\nA value of `true` means that the site supports HTTP/3 connections. A value of `false` means that it does not."
          }
        },
        "title": "HttpResponse",
        "examples": [
          {
            "hostname": "example.com",
            "http11": true,
            "http2": true,
            "http3": true
          }
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "x-examples": {
          "Error Response": {
            "error": "The hostname `https://apple.xxx` does not support ANY versions of TLS. It is probable that either the hostname is incorrect, or the website is down."
          }
        },
        "description": "Returned when there is an error.",
        "properties": {
          "error": {
            "type": "string",
            "description": "The error message."
          }
        }
      }
    },
    "responses": {
      "DomainResponse": {
        "description": "The response from a call to the `/domain` endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DomainResponse"
            },
            "examples": {
              "Example 1": {
                "value": {
                  "hostname": "example.com"
                }
              }
            }
          }
        },
        "headers": {
          "Access-Control-Allow-Credentials": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Credentials"
            }
          },
          "Access-Control-Allow-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Headers"
            }
          },
          "Access-Control-Allow-Methods": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Methods"
            }
          },
          "Access-Control-Allow-Origin": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Origin"
            }
          },
          "Access-Control-Expose-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Expose-Headers"
            }
          },
          "Access-Control-Max-Age": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Max-Age"
            }
          }
        }
      },
      "HttpResponse": {
        "description": "The response from a call to the `/http` endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/HttpResponse"
            },
            "examples": {
              "Example 1": {
                "value": {
                  "hostname": "example.com",
                  "http11": true,
                  "http2": true,
                  "http3": true
                }
              }
            }
          }
        },
        "headers": {
          "Access-Control-Allow-Credentials": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Credentials"
            }
          },
          "Access-Control-Allow-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Headers"
            }
          },
          "Access-Control-Allow-Methods": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Methods"
            }
          },
          "Access-Control-Allow-Origin": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Origin"
            }
          },
          "Access-Control-Expose-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Expose-Headers"
            }
          },
          "Access-Control-Max-Age": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Max-Age"
            }
          }
        }
      },
      "TlsResponse": {
        "description": "The response from a call to the `/tls` endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TlsResponse"
            },
            "examples": {
              "Example 1": {
                "value": {
                  "hostname": "example.com",
                  "tlsConnections": [
                    {
                      "cipherSuites": [
                        {
                          "authentication": "RSA",
                          "encryption": "AES-128-CBC",
                          "gnutlsName": "TLS_ECDHE_RSA_AES_128_CBC_SHA1",
                          "hash": "SHA-1",
                          "ianaName": "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA",
                          "isAEAD": false,
                          "isPFS": true,
                          "keyExchange": "Elliptic Curve Diffie-Hellman (Ephemeral) (ECDHE)",
                          "opensslName": "ECDHE-RSA-AES128-SHA",
                          "strength": "Weak",
                          "url": "https://ciphersuite.info/cs/TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA"
                        }
                      ],
                      "version": "TLSv1.0",
                      "versionId": 769
                    }
                  ],
                  "tlsVersions": {
                    "tls10": true,
                    "tls11": true,
                    "tls12": true,
                    "tls13": true
                  }
                }
              }
            }
          }
        },
        "headers": {
          "Access-Control-Allow-Credentials": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Credentials"
            }
          },
          "Access-Control-Allow-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Headers"
            }
          },
          "Access-Control-Allow-Methods": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Methods"
            }
          },
          "Access-Control-Allow-Origin": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Origin"
            }
          },
          "Access-Control-Expose-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Expose-Headers"
            }
          },
          "Access-Control-Max-Age": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Max-Age"
            }
          }
        }
      },
      "ErrorResponse400": {
        "description": "Bad Request: The hostname (domain) provided in the request was invalid, or the website is down.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "headers": {
          "Access-Control-Allow-Credentials": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Credentials"
            }
          },
          "Access-Control-Allow-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Headers"
            }
          },
          "Access-Control-Allow-Methods": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Methods"
            }
          },
          "Access-Control-Allow-Origin": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Origin"
            }
          },
          "Access-Control-Expose-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Expose-Headers"
            }
          },
          "Access-Control-Max-Age": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Max-Age"
            }
          }
        }
      },
      "ErrorResponse500": {
        "description": "Internal Server Error: The server experienced an error in constructing the response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "headers": {
          "Access-Control-Allow-Credentials": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Credentials"
            }
          },
          "Access-Control-Allow-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Headers"
            }
          },
          "Access-Control-Allow-Methods": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Methods"
            }
          },
          "Access-Control-Allow-Origin": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Allow-Origin"
            }
          },
          "Access-Control-Expose-Headers": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Expose-Headers"
            }
          },
          "Access-Control-Max-Age": {
            "schema": {
              "$ref": "#/components/parameters/Access-Control-Max-Age"
            }
          }
        }
      }
    },
    "requestBodies": {
      "StandardRequest": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "x-stoplight": {
                    "id": "pc5qb9w5q9yib"
                  },
                  "description": "The URL to scan."
                }
              }
            },
            "examples": {
              "Example 1": {
                "value": {
                  "url": "example.com"
                }
              }
            }
          }
        },
        "description": "The standard request input that the majority of requests accept."
      }
    },
    "parameters": {
      "StandardRequest": {
        "name": "url",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The hostname (domain) to scan."
      },
      "Access-Control-Allow-Origin": {
        "name": "Access-Control-Allow-Origin",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Standard CORS headers."
      },
      "Access-Control-Allow-Methods": {
        "name": "Access-Control-Allow-Methods",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Standard CORS headers."
      },
      "Access-Control-Allow-Headers": {
        "name": "Access-Control-Allow-Headers",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Standard CORS headers."
      },
      "Access-Control-Expose-Headers": {
        "name": "Access-Control-Expose-Headers",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Standard CORS headers."
      },
      "Access-Control-Allow-Credentials": {
        "name": "Access-Control-Allow-Credentials",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Standard CORS headers."
      },
      "Access-Control-Max-Age": {
        "name": "Access-Control-Max-Age",
        "in": "header",
        "required": true,
        "schema": {
          "type": "number"
        },
        "description": "Standard CORS headers."
      }
    }
  },
...
}